Add files via upload
Download the ZIP File and extract it to whichever folder you want. Create a new Browser Source in OBS, calling back to the downloaded folder with all these files and select index.html. If you have Cider installed and working, make sure the websocket connection is enabled in your application and once its all done, you should be able to view it.